removed -T from ln and other misc.

+++ install.sh
@@ -1,11 +1,19 @@
-#!/bin/bash
+#!/bin/sh
+set -e
 
-# submodules
 git submodule update --init
 
 DIR=`dirname $(readlink -f $0)`
-# vim config directory
-ln -sfv -T $DIR/vim ~/.vim
-# vimrc and gvimrc:
-ln -sfv $DIR/vimrc ~/.vimrc
+
+VIMDIR=$HOME/.vim
+if [ ! -L $VIMDIR -a \( -d $VIMDIR -o -f $VIMDIR \) ]; then
+  echo file $VIMDIR exists, please remove it
+  exit 1
+fi
+
+# on linux ln symlinks directories
+# in some strange manner if destination exists
+[ -L $VIMDIR ] && rm -v $VIMDIR
+ln -sfv $DIR/vim    $VIMDIR
+ln -sfv $DIR/vimrc  ~/.vimrc
 ln -sfv $DIR/gvimrc ~/.gvimrc
